// typechain-types/konsultasi.sol/ConsultationResult.ts
export declare namespace ConsultationResult {
  export type DataStruct = {
    nama: PromiseOrValue<string>;
    namaDokter: PromiseOrValue<string>;
    tanggal: PromiseOrValue<string>;
    keluhan: PromiseOrValue<string>;
    diagnosa: PromiseOrValue<string>;
    tensi: PromiseOrValue<string>;
    gula: PromiseOrValue<string>;
    wallet: PromiseOrValue<string>;
  };

  export type DataStructOutput = [
    string,
    string,
    string,
    string,
    string,
    string,
    string,
    string
  ] & {
    nama: string;
    namaDokter: string;
    tanggal: string;
    keluhan: string;
    diagnosa: string;
    tensi: string;
    gula: string;
    wallet: string;
  };
}